Fixing issues with not being able to load dlls from zipped mods.
There were two problems. First the OpenFile function in the ZipModFilesystem wasn't properly using the modRoot variable, so I fixed it to use that varirable correctly.
Second, the ModAssemblyLoadContext would return a deflated stream for zips which could not be read. to solve this, I set it up to copy it to a memory stream first in that case.