Cleaning up and simplifying things, fixing some small bugs
I noticed that this PR is removing the way it was previously separating things into tasks and loading them after. I think that's fine, but since we're doing that, it actually lets us clean things up and simplify things a little more, so now instead of adding all the assets afterwards, we're able to just do it in place while loading the assets. I also removed the LoadDirectoryRecursive function, because it didn't really feel needed after simplifying. I did have to change the order a little to accommodate this, because it means shaders must be loaded before models.
I also moved the modFs null check to the beginning, since every asset type was checking that, so it's simpler to just do that first.
I also noticed a bug where because of these changes, every mod was adding the vanilla skin. I reworked this, so the vanilla skin is now just set up when we create the vanilla mod.
I also fixed a minor bug, where I think we need to call `Language.Current.Use();` after all mods are loaded, in case any new characters are used that need to be registered.
I also changed it to show the mod name during asset load, instead of the Id. Since this is something that's displayed to the player, I think using the mod name is cleaner here.
Aside from that, I also made a few minor formatting and naming tweaks.