Changing to only load changed mods when a mod is enabled/disabled or a setting is changed.
Previously if a mod was enabled or disabled, or if a mod setting was changed that was marked as needing reload, it would always reload all mods. But because of these recent changes, we can change that to only reload mods that were actually touched, which should make that reload happen faster.
I also changed it so disabled mods don't trigger a hot reload if a file is changed, since that file isn't going to be loaded anyways.
And I also added some code to clear out hooks on reload as well.