Setting up to better handle reloading dependencies, now that we can reload single mods
Mod dependencies should now be reloaded as well when a mod is changed through hot reloading or through the settings. It also should get handled better when a mod is enabled or disabled and the dependents and dependencies get disabled/enabled as well.
It should load the dependencies similarly to how it does during initial load, so dependencies should get loaded before the mods that depend on them.
I also updated it so enabling and disabling mods should recursively affect dependencies and dependents. Previously, it only went 1 level up or down.